Fixed issues realated to imports, removed GSON, moved classes to right package
